backend/pyproject.toml
[project]
name = "repolyze-ai-backend"
version = "0.1.0"
description = "RepolyzeAI - AI-powered repository security analysis backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "pydantic-settings>=2.7.0",
    "supabase>=2.11.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",
    "weasyprint>=63.0",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]
tools = [
    "semgrep>=1.100.0",
]

[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.setuptools.packages.find]
include = ["app*", "workers*", "report*"]
